Collatz Sequence
I've created a Collatz Sequence using Python. This is a practice problem for ATBS in Chapter 3. A collatz sequence is when you input a positive number, then you do floor division ( x // 2) if it's an even number or do "3x+1" if it's an odd number. You keep doing this until your output is equal to 1.

This is also known as "The simplest impossible math problem"